+++ packages/react-container/src/components/Sentry.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import { Sentry as SentryComponent, SentryProps } from '@cognite/react-sentry';
+import { History } from 'history';
 
 import { ConditionalWrapperWithProps } from './ConditionalWrapper';
 
-export const Sentry: React.FC<SentryProps> = ({ children }) => (
-  <SentryComponent>
+export const Sentry: React.FC<SentryProps> = ({ children, history }) => (
+  <SentryComponent history={history}>
     <>{children}</>
   </SentryComponent>
 );
@@ -12,8 +13,11 @@ export const Sentry: React.FC<SentryProps> = ({ children }) => (
 export const ConditionalSentry: React.FC<{
   children: React.ReactElement;
   disabled?: boolean;
-}> = ({ disabled, children }) => (
-  <ConditionalWrapperWithProps condition={!disabled} wrap={Sentry}>
-    {children}
-  </ConditionalWrapperWithProps>
-);
+  history: History;
+}> = ({ disabled, children }) => {
+  return (
+    <ConditionalWrapperWithProps condition={!disabled} wrap={Sentry}>
+      {children}
+    </ConditionalWrapperWithProps>
+  );
+};