+++ packages/react-errors/src/index.ts
@@ -1,4 +1,5 @@
 export { default as ErrorBoundary } from './ErrorBoundary';
+export { ErrorPage } from './ErrorPage';
 export { default as ErrorId } from './ErrorId';
 export { default as ErrorWatcher } from './ErrorWatcher';
 export { ImpossibleStateError, UnreachableCaseError } from './errors';