+++ packages/react-sentry/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0 Sept 11 2021
+
+- Add support for rrweb (on by default)
+- Add prop `disableAdvancedTracking` to disable rrweb mode
+
 ## 0.1.0 Mar 9 2021
 
 ### Changes